BREAKING CHANGE: make createMap and applyDiff synchronous
Use a promise chain internally to apply changes sequentially

Also make the diff computation more robust
